STIX2 export refactored
